+++ level.go
@@ -9,28 +9,40 @@ import (
 	"strings"
 )
 
-// Level normalizes the parameter transforming it to upper case, and
-// validates it (currently, only DEBUG, INFO, WARN, and ERROR are accepted.)
-// If valid, a corresponding instance of [slog.Level] is returned as well
-// as a nil error value. If invalid, [slog.LevelInfo] is returned along
-// with a non nil error value (variable since it will contain the requested
-// level as-is).
+const (
+	LevelDebug = slog.LevelDebug
+	LevelInfo  = slog.LevelInfo
+	LevelWarn  = slog.LevelWarn
+	LevelError = slog.LevelError
+	LevelNone  = slog.Level(256)
+)
+
+// Level normalizes and validates a level string. Normalization will
+// transform the parameter into uppercase, while validation will make sure
+// that only DEBUG, INFO, WARN, ERROR, and NONE are accepted. If validation
+// succeeds, a corresponding instance of [slog.Level] will be returned and
+// a nil value for the error; otherwise, the result is [slog.LevelInfo] and
+// a non-nil error (variable since it will contain the requested level
+// as-is.)
 func Level(name string) (slog.Level, error) {
 	n := strings.ToUpper(name)
 	switch n {
 	case "DEBUG":
-		return slog.LevelDebug, nil
+		return LevelDebug, nil
 
 	case "INFO":
-		return slog.LevelInfo, nil
+		return LevelInfo, nil
 
 	case "WARN":
-		return slog.LevelWarn, nil
+		return LevelWarn, nil
 
 	case "ERROR":
-		return slog.LevelError, nil
+		return LevelError, nil
+
+	case "NONE":
+		return LevelNone, nil
 
 	default:
-		return slog.LevelInfo, fmt.Errorf("invalid log level: %s", name)
+		return LevelInfo, fmt.Errorf("invalid log level: %s", name)
 	}
 }