+++ logger.go
@@ -12,11 +12,12 @@ import (
 // New creates an instance of [slog.Logger] that sends its output to the
 // given writer, using the given format, and with the specified level.
 // Output lines include the event timestamp if the given parameter is set.
+// If either an invalid format or level is requested, a nil is returned
+// as well as an error (variable since it will contain the root cause.)
+//
+// Completely silent behavior can be achieved either by using a nil writer,
+// [io.Discard], or by requesting the NONE level.
 func New(output io.Writer, format string, level string, timestamps bool) (*slog.Logger, error) {
-	if output == nil {
-		output = io.Discard
-	}
-
 	f, err := Format(format)
 	if err != nil {
 		return nil, fmt.Errorf("can not create logger: %s", err)
@@ -24,6 +25,12 @@ func New(output io.Writer, format string, level string, timestamps bool) (*slog.
 	l, err := Level(level)
 	if err != nil {
 		return nil, fmt.Errorf("can not create logger: %s", err)
+	}
+
+	// If no output is requested, then the discard handler is used, which
+	// does not need to be further configured.
+	if output == nil || output == io.Discard || l == LevelNone {
+		return slog.New(slog.DiscardHandler), nil
 	}
 
 	opts := &slog.HandlerOptions{